edit type assertion
- use standalone example file
- fix missing dependencies
- using simple Color interface without full specs
- add relevant links
- include full section example
